Delete Show Update
For some reason it fucks up the page it's currently on so I made a new page to for forward it too and then redirect it back from there.

The commit seems larger because I needed JQuery to make parts of the page work. It works fine now and lastly needs styling which i feel might end up being a huge pain in the ass.