DeleteMovie now shows a dropdown and Schema Update
So I made foreign key costraints and stuff for the schema.

DeleteMovie now shows a drop-down list instead of taking in text input and Delete is now cascaded

Four screens made now and each linked to bookings.
Hopefully this will be the last schema update.
The Admin Skeleton is done,  only the link page remains which will be easy enough as a skeleton so We'll add it in while stylizing it.