SelectTime is fucntional and Back Complete
We are now 90% done with the middle layers. The Report thing he wanted is left.
The commits following this should hopefully be the last additions to our backbone.
Assuming we finish seat map tomorrow,(doubful) we have no time to waste on CSS.
We'll look at it tomorrow anyhow. Got done before midnight and also commented the problems I had in  SeatMapGen