ONLY CHECKBOX CSS APPLIED
the rest of the page still needs formatting. I say, we keep the white background as is and maybe add more seats to the place.

For the checkbox, i tried using plain css3 but, it didn't work. It would apply the property but wouldn't let me select it without ruining how it looked.

This method, works I found an open source CSS that helped me out. It's pretty fucking great and once we have a functioning project, we can strip it down.